requirements.txt
fastapi
uvicorn[standard]
python-dotenv
line-bot-sdk
requests